Updated MacTeX Recipe to grab version from website
This update will pull the version of MacTeX from their website as it
will always import into Munki as version 1.0.  On their website they
state that “The current distribution is MacTeX-(Version Here)” so I am
pulling that each time there is an update as it seems to be updated on
a yearly basis.